additions for file loading

+++ load_freq.py
@@ -1,13 +1,10 @@
 import os
 import numpy as np
-import cupy as cp
 import mne
 
-from sklearn import config_context
 from ieeg.io import get_data, DataLoader
 from analysis.grouping import group_elecs
 from ieeg.arrays.label import LabeledArray, combine, Labels
-from analysis.utils.plotting import plot_horizontal_bars
 from ieeg.decoding.decode import Decoder, get_scores, plot_all_scores
 
 fpath = os.path.expanduser("~/Box/CoganLab")
@@ -27,14 +24,14 @@ def load_data(datatype: str, out_type: type | str = float,
     loader = DataLoader(layout, conds, datatype, average, 'stats_freq',
                        '.h5')
     zscore = loader.load_dict(dtype=out_type, n_jobs=n_jobs)
-    if average:
-        zscore_ave = combine(zscore, (0, 2))
-        for key in zscore_ave.keys():
-            for k in zscore_ave[key].keys():
-                for f in zscore_ave[key][k].keys():
-                    zscore_ave[key][k][f] = zscore_ave[key][k][f][..., :200]
-    else:
-        zscore_ave = combine(zscore, (0, 3))
+    # if average:
+    #     zscore_ave = combine(zscore, (0, 2))
+    #     for key in zscore_ave.keys():
+    #         for k in zscore_ave[key].keys():
+    #             for f in zscore_ave[key][k].keys():
+    #                 zscore_ave[key][k][f] = zscore_ave[key][k][f][..., :200]
+    # else:
+    zscore_ave = combine(zscore, (0, 3))
 
     del zscore
     return LabeledArray.from_dict(zscore_ave, dtype=out_type)
@@ -48,25 +45,33 @@ def name_from_idx(idx: list[int], chs: Labels):
     return [f"D{int(p[0][1:])}-{p[1]}" for p in
              (s.split("-") for s in chs[idx])]
 
-# loader = DataLoader(layout, conds, "zscore", False, 'stats_freq',
-#                     '.h5')
-# zscore = loader.load_dict(dtype="float16")
-# zscore2 = combine(zscore, (0, 3))
-# zscore = LabeledArray.from_dict(zscore2, dtype="float16")
-if not os.path.exists("zscores.npy"):
-    zscores = load_data("zscore", "float16", False, 12)
-    zscores.tofile("zscores")
-zscores = LabeledArray.fromfile("zscores", mmap_mode='r')
-
-sigs = load_data("significance", out_type=bool)
+conds_all = {"resp": (-1, 1), "aud_ls": (-0.5, 1.5),
+                 "aud_lm": (-0.5, 1.5), "aud_jl": (-0.5, 1.5),
+                 "go_ls": (-0.5, 1.5), "go_lm": (-0.5, 1.5),
+                 "go_jl": (-0.5, 1.5)}
+loader = DataLoader(layout, conds_all, 'significance', True, 'stats_freq',
+                   '.h5')
+filemask = os.path.join(layout.root, 'derivatives', 'stats_freq', 'combined', 'mask')
+if not os.path.exists(filemask + ".npy"):
+    sigs = LabeledArray.from_dict(combine(loader.load_dict(
+        dtype=bool, n_jobs=-1), (0, 2)), dtype=bool)
+    sigs.tofile(filemask)
+else:
+    sigs = LabeledArray.fromfile(filemask)
 
 AUD, SM, PROD, sig_chans = group_elecs(sigs, sigs.labels[1], sigs.labels[0])
+idxs = {'SM': SM, 'AUD': AUD, 'PROD': PROD, 'sig_chans': sig_chans}
+
+filez = os.path.join(layout.root, 'derivatives', 'stats_freq', 'combined', 'zscores')
+if not os.path.exists(filez + ".npy"):
+    zscores = load_data("zscore", "float16", False, -1)
+    zscores.tofile(filez)
+zscores = LabeledArray.fromfile(filez, mmap_mode='r')
 
 SM = sorted(SM)
 AUD = sorted(AUD)
 PROD = sorted(PROD)
 sig_chans = sorted(sig_chans)
-avg = None
 
 # %% plot data
 do_plot=False
@@ -79,7 +84,10 @@ if do_plot:
     picked = mne.time_frequency.EpochsTFRArray(
         info, avg.__array__(), times, avg.labels[2],
         events=events, event_id=event_id, drop_log=tuple(() for _ in range(7)))
-    picked['aud_ls'].pick(AUD).average().plot(0)
+    x = picked['aud_ls'].pick(sorted(SM)).average()
+    mne.time_frequency.AverageTFRArray(x.copy().pick(0).info,
+                                       np.mean(x.data, axis=0, keepdims=True),
+                                       x.times, x.freqs).plot(0)
 
 # %% plot brain
 plot_brain = False
@@ -92,7 +100,7 @@ if plot_brain:
         br = plot_on_average(subjects, picks=picks, hemi='both', color=[rgb]*len(idx), fig=br)
 
 # %% word decoding
-decode = True
+decode = False
 if decode:
     decoder = Decoder({'heat': 0, 'hoot': 1, 'hot': 2, 'hut': 3},
                       5, 10, 1, 'train', explained_variance=0.8, da_type='lda')
@@ -105,12 +113,15 @@ if decode:
     conds = [['aud_ls', 'aud_lm'], ['go_ls', 'go_lm'], 'resp']
     colors = [[0, 0, 1], [1, 0, 0], [0, 1, 0], [0.5, 0.5, 0.5]]
 
-    for values in get_scores(zscores, decoder, idxs, conds,
-                                  names, on_gpu=False, shuffle=False, **window_kwargs):
-        key = decoder.current_job
-        scores_dict[key] = values
+    if not os.path.exists("true_scores_freq.npz"):
+        for values in get_scores(zscores, decoder, idxs, conds,
+                                 names, on_gpu=True, shuffle=False, **window_kwargs):
+            key = decoder.current_job
+            scores_dict[key] = values
 
-    np.savez('true_scores_freq', **scores_dict)
+        np.savez('true_scores_freq', **scores_dict)
+    else:
+        scores_dict = dict(np.load("true_scores_freq.npz", allow_pickle=True))
 
     plots = {}
     for key, values in scores_dict.items():
@@ -123,10 +134,12 @@ if decode:
 
     decoder = Decoder({'heat': 0, 'hoot': 1, 'hot': 2, 'hut': 3},
                       5, 50, 1, 'train', explained_variance=0.8, da_type='lda')
-    scores_dict2 = {}
-    for values in get_scores(zscores, decoder, idxs, conds,
-                                  names, on_gpu=False, shuffle=True, **window_kwargs):
-        key = decoder.current_job
-        scores_dict2[key] = values
+    if not os.path.exists("shuffle_scores_freq.npz"):
+        for values in get_scores(zscores, decoder, idxs, conds,
+                                 names, on_gpu=True, shuffle=True, **window_kwargs):
+            key = decoder.current_job
+            scores_dict2[key] = values
 
-    np.savez('shuffle_scores_freq', **scores_dict2)
+        np.savez('shuffle_scores_freq', **scores_dict2)
+    else:
+        scores_dict2 = dict(np.load("shuffle_scores_freq.npy", allow_pickle=True))
